api/pyproject.toml
[project]
name = "rayai-platform-api"
version = "1.0.0"
description = "FastAPI backend for RayAI Platform sandbox orchestration"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "ray>=2.49.2",
    "ray-agents @ git+https://github.com/ray-project/ray-agents.git",
    "psycopg2-binary>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "pytest>=7.0.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
